Flyttade koden från homePage till home
