+++ netflix/src/App.jsx
@@ -1,11 +1,10 @@
 import { useState } from "react";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import Home from "./views/Home/Home";
 import FilmView from "./views/FilmView/FilmView";
 import Favourites from "./views/Favourites/Favourites";
 import Categories from "./views/Categories/Categories";
+import Home from "./views/Home/Home";
 import "./App.css";
-import HomePage from "./components/Home/HomePage";
 
 const router = createBrowserRouter([
   {
@@ -30,7 +29,6 @@ function App() {
   return (
     <div className="App">
       <RouterProvider router={router} />
-      <HomePage />
     </div>
   );
 }
