--- netflix/src/components/Home/homePage.css
@@ -1,4 +1,0 @@
-img {
-  max-width: 500px;
-  max-height: 500px;
-}
