+++ netflix/src/views/Home/Home.css
@@ -1,0 +1,4 @@
+img {
+  max-width: 500px;
+  max-height: 500px;
+}
